// src/components/sort-card/types.ts
export interface IfilterParamIcons {
    "": React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
    "isHot": React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
    "isSpicy": React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
    "isNew": React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
    "isTop": React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
    
  }

export type filterParams = "" | "isHot" | "isSpicy" | "isNew" | "isTop"